requirements.txt
# requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=1.22.0
requests>=2.26.0
python-dotenv>=0.19.0
together>=1.0.0
pydantic>=1.8.0